refactor: replacing value objects in the entity with simplified variables
feat: add two endpoints, get vehicles by vin and by plate

+++ src/main/java/com/services/vehicle/infrastructure/web/dto/VehicleControllerRequestDTO.java
@@ -13,6 +13,6 @@ public record VehicleControllerRequestDTO(
         String bodyType,
         String fuelType,
         String engineNumber,
-        Integer initialKm,
-        Integer currentKm
+        Double initialKm,
+        Double currentKm
 ) {}
